fix(parseMessage): Factor parseMessage to another file for decoupling.
