+++ lib/irc.js
@@ -24,9 +24,8 @@ var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 
 var colors = require('./colors');
+var parseMessage = require('./parse_message');
 exports.colors = colors;
-
-var replyFor = require('./codes');
 
 var lineDelimiter = new RegExp('\r\n|\r|\n')
 
@@ -1026,68 +1025,3 @@ Client.prototype._updateMaxLineLength = function() {
     // target is determined in _speak() and subtracted there
     this.maxLineLength = 497 - this.nick.length - this.hostMask.length;
 };
-/*
- * parseMessage(line, stripColors)
- *
- * takes a raw "line" from the IRC server and turns it into an object with
- * useful keys
- */
-function parseMessage(line, stripColors) {
-    var message = {};
-    var match;
-
-    if (stripColors) {
-        line = line.replace(/[\x02\x1f\x16\x0f]|\x03\d{0,2}(?:,\d{0,2})?/g, '');
-    }
-
-    // Parse prefix
-    match = line.match(/^:([^ ]+) +/);
-    if (match) {
-        message.prefix = match[1];
-        line = line.replace(/^:[^ ]+ +/, '');
-        match = message.prefix.match(/^([_a-zA-Z0-9\[\]\\`^{}|-]*)(!([^@]+)@(.*))?$/);
-        if (match) {
-            message.nick = match[1];
-            message.user = match[3];
-            message.host = match[4];
-        }
-        else {
-            message.server = message.prefix;
-        }
-    }
-
-    // Parse command
-    match = line.match(/^([^ ]+) */);
-    message.command = match[1];
-    message.rawCommand = match[1];
-    message.commandType = 'normal';
-    line = line.replace(/^[^ ]+ +/, '');
-
-    if (replyFor[message.rawCommand]) {
-        message.command     = replyFor[message.rawCommand].name;
-        message.commandType = replyFor[message.rawCommand].type;
-    }
-
-    message.args = [];
-    var middle, trailing;
-
-    // Parse parameters
-    if (line.search(/^:|\s+:/) != -1) {
-        match = line.match(/(.*?)(?:^:|\s+:)(.*)/);
-        middle = match[1].trimRight();
-        trailing = match[2];
-    }
-    else {
-        middle = line;
-    }
-
-    if (middle.length)
-        message.args = middle.split(/ +/);
-
-    if (typeof (trailing) != 'undefined' && trailing.length)
-        message.args.push(trailing);
-
-    return message;
-}
-
-exports.parseMessage = parseMessage;
